+++ packages/components/molecules/f-tabs/test-utils/component-objects/f-tabs.component.js
@@ -2,6 +2,7 @@ const Page = require('@justeat/f-wdio-utils/src/page.object');
 
 module.exports = class Tabs extends Page {
     get component () { return $('[data-test-id="tabs-component"]'); }
+    get tabButtons () { return $$('[data-test-id*="tab-button"]'); }
 
     open () {
         super.openComponent('molecule', 'vue-tabs-component');
@@ -14,4 +15,14 @@ module.exports = class Tabs extends Page {
     isComponentDisplayed () {
         return this.component.isDisplayed();
     }
-}
+
+    isTabButtonDisplayed () {
+        return this.tabButton.isDisplayed();
+    }
+
+    set expectedTabButton (tabTitle) {
+        this.tabButtonValue = this.tabButtons.filter(element => element.getText().includes(tabTitle))[0];
+    }
+
+    get tabButton () { return this.tabButtonValue != null ? this.tabButtonValue : 'Please set an expected tab button value'; }
+};