+++ packages/components/atoms/f-error-message/CHANGELOG.md
@@ -3,8 +3,13 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-Latest (add to next release)
+v0.4.0
 ------------------------------
+*March 3, 2021*
+
+### Added
+- Added component test file and tests
+
 *February 25, 2021*
 
 ### Changed