+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -3,6 +3,16 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.64.0
+------------------------------
+*March 4, 2021*
+
+### Changed
+- `analytics` to `fCheckout.module`.
+- `analytics.module` to `fCheckoutAnalytics.module`.
+
+### Removed
+- `f-trak` dependency.
 
 v0.63.0
 ------------------------------